\name{greenwood.table}
\alias{greenwood.table}
\docType{data}
\title{ Upper percentage points for Greenwood statistic}
\description{
	Upper percentage points for Greenwood statistic.
}

\usage{data(greenwood.table)}
\format{
  A data frame of 23 observations on the following 5 variables:
  \describe{
    \item{\code{n}}{sample size}
    \item{\code{q10}}{quantile at significance level 0.1}
    \item{\code{q5}}{quantile at significance level 0.05}
    \item{\code{q2.5}}{quantile at significance level 0.025}
    \item{\code{q1}}{quantile at significance level 0.01}
  }
}

\source{
	D Agostino R.B., Stephens M.A. (1986), Goodness-of-fit techniques, CRC Press, New York, page 340, table 8.3.
}

\keyword{datasets}
